requirements.txt
torch>=2.0,<3.0
torchvision>=0.15
timm>=0.9
scikit-learn>=1.3
pillow>=9.4
matplotlib>=3.7
tqdm>=4.66
pandas>=1.5
openpyxl>=3.1
numpy>=1.23
streamlit>=1.49.1